image upload fix

+++ src/store/modules/teams.js
@@ -96,10 +96,10 @@ export default {
             const ext = payload.ext
             let imageUrl
             const imagePath = fb.storage().ref('users').child(user).child('teams').child(`${payload.teamId}${ext}`)
-            imagePath.delete()
-                .then(() => {
-                    return imagePath.put(payload.image)
-                })
+            if (fb.storage().ref('users').child(user).child('teams').child(`${payload.teamId}${ext}`)) {
+                fb.storage().ref('users').child(user).child('teams').child(`${payload.teamId}${ext}`).delete()
+            }
+            imagePath.put(payload.image)
                 .then(() => {
                     return imagePath.getDownloadURL()
                 })
@@ -119,7 +119,9 @@ export default {
         deleteTeam({ commit, rootState }, payload) {
             commit('SET_LOADING', true, { root: true })
             const user = rootState.user.user.id
-            fb.storage().ref('users').child(user).child('teams').child(`${payload.teamId}${payload.ext}`).delete()
+            if (fb.storage().ref('users').child(user).child('teams').child(`${payload.teamId}${payload.ext}`)) {
+                fb.storage().ref('users').child(user).child('teams').child(`${payload.teamId}${payload.ext}`).delete()
+            }
             fb.database().ref('users').child(user).child('games').child(payload.gameId).child('teams').child(payload.teamId).remove()
                 .then(() => {
                     commit("DELETE_TEAM", payload)
